+++ app/src/main/java/com/nagane/table/data/model/Menu.kt
@@ -1,19 +1,18 @@
 package com.nagane.table.data.model
 
 data class Menu(
-    val menuNo : Int,
+    val menuNo : Long,
     val menuName: String,
     val price : Int,
     val soldOut : Boolean
 )
-
 
 data class MenuData(
     val menuList: List<Menu>
 )
 
 data class MenuDetail(
-    val menuNo : Int,
+    val menuNo : Long,
     val menuName: String,
     val price : Int,
     val menuImage: String,
